+++ packages/sst/src/constructs/SsrSite.ts
@@ -208,10 +208,10 @@ export interface SsrSiteProps {
    * @default nodejs18.x
    * @example
    * ```js
-   * runtime: "nodejs16.x",
+   * runtime: "nodejs20.x",
    * ```
    */
-  runtime?: "nodejs16.x" | "nodejs18.x";
+  runtime?: "nodejs16.x" | "nodejs18.x" | "nodejs20.x";
   /**
    * The instruction set [architecture](https://docs.aws.amazon.com/lambda/latest/dg/foundation-arch.html) for the SSR function.
    *