+++ packages/sst/test/constructs/Function.test.ts
@@ -218,16 +218,16 @@ test("copyFiles nonexistent", async () => {
   }).rejects.toThrow(/no such file/);
 });
 
-test("runtime: nodejs18.x", async () => {
+test("runtime: nodejs20.x", async () => {
   const app = await createApp();
   const stack = new Stack(app, "stack");
   new Function(stack, "Function", {
     handler: "test/constructs/lambda.handler",
-    runtime: "nodejs18.x",
+    runtime: "nodejs20.x",
   });
   await app.finish();
   hasResource(stack, "AWS::Lambda::Function", {
-    Runtime: "nodejs18.x",
+    Runtime: "nodejs20.x",
   });
 });
 
@@ -757,6 +757,34 @@ test("url.cors: allowMethods *", async () => {
     Cors: {
       AllowMethods: ["*"],
     },
+  });
+});
+
+test("url.streaming: true", async () => {
+  const stack = new Stack(await createApp(), "stack");
+  const fn = new Function(stack, "Function", {
+    handler: "test/lambda.handler",
+    url: {
+      streaming: true,
+    },
+  });
+  expect(fn.url).toBeDefined();
+  hasResource(stack, "AWS::Lambda::Url", {
+    InvokeMode: lambda.InvokeMode.RESPONSE_STREAM,
+  });
+});
+
+test("url.streaming: false", async () => {
+  const stack = new Stack(await createApp(), "stack");
+  const fn = new Function(stack, "Function", {
+    handler: "test/lambda.handler",
+    url: {
+      streaming: false,
+    },
+  });
+  expect(fn.url).toBeDefined();
+  hasResource(stack, "AWS::Lambda::Url", {
+    InvokeMode: lambda.InvokeMode.BUFFERED,
   });
 });