+++ packages/create-sst/CHANGELOG.md
@@ -1,5 +1,15 @@
 # create-sst
 
+## 2.36.4
+
+### Patch Changes
+
+- [#3528](https://github.com/sst/sst/pull/3528) [`4732d21ec`](https://github.com/sst/sst/commit/4732d21eca6ef1911ba2ec970b916f3d2ebf46fd) Thanks [@fwang](https://github.com/fwang)! - Update CDK to 2.110.1
+
+- [`6bcf40c13`](https://github.com/sst/sst/commit/6bcf40c1399bdc4093d324df52e5f2a939037ef7) Thanks [@fwang](https://github.com/fwang)! - Update constructs to 10.3.0
+
+## 2.36.3
+
 ## 2.36.2
 
 ### Patch Changes